refactor: Split settings.js
Move the page logics to separate files
 - account types
 - import formats
 - planning cycle

fix: Test errors and warnings due to renaming
fix: Too many function parameter in insert_ignore of the planning cycle repository
fix: Loading the categories file/methods
