--- docker-compose.override.yml.example
@@ -1,18 +1,0 @@
-# Docker Compose Override Template (tested on Synology Container Manager)
-#
-
-services:
-  api:
-    # Override ports if conflict with other containers
-    # Format: "HOST_PORT:CONTAINER_PORT"
-    ports:
-      - "8000:8000"  # Change first number to avoid conflicts (e.g., "8001:8000")
-    
-    # Override volume mount paths for Synology
-    # Use absolute Synology paths: /volume1/docker/FiniA/cfg
-    volumes:
-      - ./cfg:/app/cfg:ro  # Synology absolute path
-    
-    # Environment overrides (optional)
-    # environment:
-    #   - LOG_LEVEL=debug
