+++ src/web/categories.html
@@ -336,6 +336,7 @@
   <script src="help_documents.js"></script>
   <script src="help.js"></script>
   <script src="app.js"></script>
+
   <script>
     let allCategories = [];
     let currentEditingId = null;
